LPD-82612 Add check permission

+++ modules/dxp/apps/saml/saml-web/src/main/java/com/liferay/saml/web/internal/portlet/action/DeleteSamlSpIdpConnectionMVCActionCommand.java
@@ -9,7 +9,9 @@ import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCActionCommand;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCActionCommand;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.saml.constants.SamlPortletKeys;
+import com.liferay.saml.persistence.model.SamlSpIdpConnection;
 import com.liferay.saml.persistence.service.SamlSpIdpConnectionLocalService;
+import com.liferay.saml.web.internal.util.SamlPermissionUtil;
 
 import javax.portlet.ActionRequest;
 import javax.portlet.ActionResponse;
@@ -35,8 +37,17 @@ public class DeleteSamlSpIdpConnectionMVCActionCommand
 			ActionRequest actionRequest, ActionResponse actionResponse)
 		throws Exception {
 
+		long samlSpIdpConnectionId = ParamUtil.getLong(
+			actionRequest, "samlSpIdpConnectionId");
+
+		SamlSpIdpConnection samlSpIdpConnection =
+			_samlSpIdpConnectionLocalService.getSamlSpIdpConnection(
+				samlSpIdpConnectionId);
+
+		SamlPermissionUtil.checkPermission(samlSpIdpConnection.getCompanyId());
+
 		_samlSpIdpConnectionLocalService.deleteSamlSpIdpConnection(
-			ParamUtil.getLong(actionRequest, "samlSpIdpConnectionId"));
+			samlSpIdpConnectionId);
 	}
 
 	@Reference
